Repository 패턴 적용
