+++ app/src/androidTest/java/campus/tech/kakao/map/uiTest/SearchActivityTest.kt
@@ -8,7 +8,7 @@ import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import campus.tech.kakao.map.R
-import campus.tech.kakao.map.view.SearchActivity
+import campus.tech.kakao.map.presentation.view.SearchActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
