+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
             android:value="@string/KAKAO_API_KEY" />
 
         <activity
-            android:name=".view.SplashActivity"
+            android:name=".presentation.view.SplashActivity"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"
             android:exported="true">
             <intent-filter>
@@ -31,10 +31,10 @@
         </activity>
 
         <activity android:name=".MainActivity" android:exported="true" />
-        <activity android:name=".view.SearchActivity" />
+        <activity android:name=".presentation.view.SearchActivity" />
 
         <service
-            android:name=".service.MyFirebaseMessagingService"
+            android:name=".presentation.view.service.MyFirebaseMessagingService"
             android:exported="true">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
