+++ app/src/main/java/campus/tech/kakao/map/data/entity/KeywordEntity.kt
@@ -1,0 +1,11 @@
+package campus.tech.kakao.map.data.entity
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "keywords")
+data class KeywordEntity(
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @ColumnInfo(name = "keyword") val keyword: String
+)
