--- app/src/main/java/campus/tech/kakao/map/database/AppDatabase.kt
@@ -1,14 +1,0 @@
-package campus.tech.kakao.map.database
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-import campus.tech.kakao.map.entity.KeywordEntity
-import campus.tech.kakao.map.entity.LocationEntity
-import campus.tech.kakao.map.repository.keyword.KeywordDao
-import campus.tech.kakao.map.repository.location.LocationDao
-
-@Database(entities = [KeywordEntity::class, LocationEntity::class], version = 2)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun keywordDao(): KeywordDao
-    abstract fun locationDao(): LocationDao
-}
