--- app/src/main/java/campus/tech/kakao/map/di/AppModule.kt
@@ -1,30 +1,0 @@
-package campus.tech.kakao.map.di
-
-import campus.tech.kakao.map.api.KakaoLocalApi
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AppModule {
-
-    @Provides
-    @Singleton
-    fun provideRetrofit(): Retrofit {
-        return Retrofit.Builder()
-            .baseUrl("https://dapi.kakao.com")
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    @Provides
-    @Singleton
-    fun provideKakaoLocalApi(retrofit: Retrofit): KakaoLocalApi {
-        return retrofit.create(KakaoLocalApi::class.java)
-    }
-}
