+++ app/src/main/java/campus/tech/kakao/map/domain/model/Location.kt
@@ -1,0 +1,9 @@
+package campus.tech.kakao.map.domain.model
+
+data class Location(
+    val place: String,
+    val address: String,
+    val category: String,
+    val latitude: Double,
+    val longitude: Double
+)
