+++ app/src/main/java/campus/tech/kakao/map/domain/repository/keyword/KeywordDao.kt
@@ -1,0 +1,20 @@
+package campus.tech.kakao.map.domain.repository.keyword
+
+import androidx.room.Dao
+import androidx.room.Delete
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
+import campus.tech.kakao.map.data.entity.KeywordEntity
+
+@Dao
+interface KeywordDao {
+    @Query("SELECT * FROM keywords")
+    suspend fun getAllKeywords(): List<KeywordEntity>
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertKeyword(keyword: KeywordEntity)
+
+    @Delete
+    suspend fun deleteKeyword(keyword: KeywordEntity)
+}
