+++ app/src/main/java/campus/tech/kakao/map/domain/repository/module/AppModule.kt
@@ -1,0 +1,30 @@
+package campus.tech.kakao.map.domain.repository.module
+
+import campus.tech.kakao.map.data.api.KakaoLocalApi
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(): Retrofit {
+        return Retrofit.Builder()
+            .baseUrl("https://dapi.kakao.com")
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideKakaoLocalApi(retrofit: Retrofit): KakaoLocalApi {
+        return retrofit.create(KakaoLocalApi::class.java)
+    }
+}
