--- app/src/main/java/campus/tech/kakao/map/model/Location.kt
@@ -1,9 +1,0 @@
-package campus.tech.kakao.map.model
-
-data class Location(
-    val place: String,
-    val address: String,
-    val category: String,
-    val latitude: Double,
-    val longitude: Double
-)
