+++ app/src/main/java/campus/tech/kakao/map/presentation/view/OnKeywordItemClickListener.kt
@@ -1,0 +1,6 @@
+package campus.tech.kakao.map.presentation.view
+
+interface OnKeywordItemClickListener {
+    fun onKeywordItemDeleteClick(keyword: String)
+    fun onKeywordItemClick(keyword: String)
+}
