+++ app/src/main/java/campus/tech/kakao/map/presentation/view/OnSearchItemClickListener.kt
@@ -1,0 +1,7 @@
+package campus.tech.kakao.map.presentation.view
+
+import campus.tech.kakao.map.domain.model.Location
+
+interface OnSearchItemClickListener {
+    fun onSearchItemClick(location: Location)
+}
