--- app/src/main/java/campus/tech/kakao/map/repository/keyword/KeywordDao.kt
@@ -1,20 +1,0 @@
-package campus.tech.kakao.map.repository.keyword
-
-import androidx.room.Dao
-import androidx.room.Delete
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import campus.tech.kakao.map.entity.KeywordEntity
-
-@Dao
-interface KeywordDao {
-    @Query("SELECT * FROM keywords")
-    suspend fun getAllKeywords(): List<KeywordEntity>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertKeyword(keyword: KeywordEntity)
-
-    @Delete
-    suspend fun deleteKeyword(keyword: KeywordEntity)
-}
