--- app/src/main/java/campus/tech/kakao/map/repository/location/LocationDao.kt
@@ -1,11 +1,0 @@
-package campus.tech.kakao.map.repository.location
-
-import androidx.room.Dao
-import androidx.room.Query
-import campus.tech.kakao.map.entity.LocationEntity
-
-@Dao
-interface LocationDao {
-    @Query("SELECT * FROM location WHERE category_group_name = :category")
-    suspend fun searchByCategory(category: String): List<LocationEntity>
-}
