--- app/src/main/java/campus/tech/kakao/map/repository/location/LocationSearcher.kt
@@ -1,22 +1,0 @@
-package campus.tech.kakao.map.repository.location
-
-import campus.tech.kakao.map.model.Location
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class LocationSearcher @Inject constructor(private val locationDao: LocationDao) {
-
-    suspend fun search(keyword: String): List<Location> {
-        val locationEntities = locationDao.searchByCategory(keyword)
-        return locationEntities.map {
-            Location(
-                place = it.placeName,
-                address = it.addressName,
-                category = it.categoryGroupName,
-                latitude = it.latitude,
-                longitude = it.longitude
-            )
-        }
-    }
-}
