--- app/src/main/java/campus/tech/kakao/map/view/OnKeywordItemClickListener.kt
@@ -1,6 +1,0 @@
-package campus.tech.kakao.map.view
-
-interface OnKeywordItemClickListener {
-    fun onKeywordItemDeleteClick(keyword: String)
-    fun onKeywordItemClick(keyword: String)
-}
