--- app/src/main/java/campus/tech/kakao/map/view/OnSearchItemClickListener.kt
@@ -1,7 +1,0 @@
-package campus.tech.kakao.map.view
-
-import campus.tech.kakao.map.model.Location
-
-interface OnSearchItemClickListener {
-    fun onSearchItemClick(location: Location)
-}
